requirements.txt

Flask==0.12.2
python-dotenv==0.10.3
flask-limiter==1.0.1
flask-debugtoolbar==0.10.1
Flask-SQLAlchemy==2.4.0
